Initial Commit: "Deja Vu"
Copy of "Deja Vu", the text-based choose-your-own-adventure game that I submit as my project for the Intro to Programming Course through Udacity. You awake in a mysterious place, devoid of memory. Find out where, and who, you are as you explore and learn the mysteries that all of us may someday have to face.